Separated icon props function
